+++ src/main/java/mk/ukim/finki/wpproekt/model/exceptions/KorisnikNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(code = HttpStatus.NOT_FOUND)
 public class KorisnikNotFoundException extends RuntimeException {
 
-    public KorisnikNotFoundException(Integer id) {
+    public KorisnikNotFoundException(Long id) {
         super(String.format("User with id: %d was not found", id));
     }
 }
