+++ src/main/java/mk/ukim/finki/wpproekt/model/exceptions/LekAlreadyInShoppingCartException.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.PRECONDITION_FAILED)
 public class LekAlreadyInShoppingCartException extends RuntimeException {
 
-    public LekAlreadyInShoppingCartException(Integer id, String username) {
+    public LekAlreadyInShoppingCartException(Long id, String username) {
         super(String.format("Drug with id: %d already exists in shopping cart for user with username %s", id, username));
     }
 }
