+++ src/main/java/mk/ukim/finki/wpproekt/model/exceptions/LektNotFoundException.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class LektNotFoundException extends RuntimeException {
 
-    public LektNotFoundException(Integer id) {
+    public LektNotFoundException(Long id) {
         super(String.format("Drug with id: %d was not found", id));
     }
 }
