+++ src/main/java/mk/ukim/finki/wpproekt/repository/impl/TerminRepositoryImpl.java
@@ -18,19 +18,19 @@ public class TerminRepositoryImpl implements TerminRepositoryCustom {
     private EntityManager entityManager;
 
     @Override
-    public List<Termin> findAllByTerminId(Integer termin_id) {
+    public List<Termin> findAllByTerminId(Long termin_id) {
         return entityManager.createQuery("select o from Termin o where o.termin_id = :termin_id",
                 Termin.class).setParameter("termin_id", termin_id).getResultList();
     }
 
     @Override
-    public Termin findByTerminId(Integer termin_id) {
+    public Termin findByTerminId(Long termin_id) {
         return entityManager.createQuery("select o from Termin o where o.termin_id = :termin_id",
                 Termin.class).setParameter("termin_id", termin_id).setMaxResults(1).getSingleResult();
     }
 
     @Override
-    public void deleteByTerminId(Integer termin_id) {
+    public void deleteByTerminId(Long termin_id) {
         entityManager.createQuery("delete from Termin o where o.termin_id = :termin_id");
     }
