+++ src/main/java/mk/ukim/finki/wpproekt/sevice/RezervacijaService.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 
 public interface RezervacijaService {
     void save (Rezervacija rezervacija);
-    Optional<Rezervacija> findById (Integer id);
-    List<Rezervacija> findAllValidReservations(Integer pacient_id);
-    List<Rezervacija> findAllValidReservationsForDoktor(Integer doktor_id);
+    Optional<Rezervacija> findById (Long id);
+    List<Rezervacija> findAllValidReservations(Long pacient_id);
+    List<Rezervacija> findAllValidReservationsForDoktor(Long doktor_id);
 
 }
